Fixed compile error if ImageMagick has double as Quantum type (closes #140)

+++ bitmap.c
@@ -241,7 +241,7 @@ bool cText2SkinBitmap::LoadMagick(const char *Filename, int height, int width, i
 			const PixelPacket *pix = (*it).getConstPixels(0, 0, w, h);
 			for (int iy = 0; iy < h; ++iy) {
 				for (int ix = 0; ix < w; ++ix) {
-					tColor col = (~(pix->opacity * 255 / MaxRGB) << 24) | ((pix->red * 255 / MaxRGB) << 16) | ((pix->green * 255 / MaxRGB) << 8) | (pix->blue * 255 / MaxRGB);
+					tColor col = (~int(pix->opacity * 255 / MaxRGB) << 24) | (int(pix->red * 255 / MaxRGB) << 16) | (int(pix->green * 255 / MaxRGB) << 8) | int(pix->blue * 255 / MaxRGB);
 					bmp->DrawPixel(ix, iy, col);
 					++pix;
 				}
